Develop

+++ src/main/java/com/forums/controller/CommentController.java
@@ -1,5 +1,8 @@
 package com.forums.controller;
 
+import static org.springframework.http.ResponseEntity.badRequest;
+import static org.springframework.http.ResponseEntity.ok;
+
 import com.forums.model.Comment;
 import com.forums.model.Post;
 import com.forums.repository.AccountRepository;
@@ -7,6 +10,7 @@ import com.forums.repository.CommentRepository;
 import com.forums.repository.PostRepository;
 import java.util.ArrayList;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -45,18 +49,33 @@ public class CommentController {
 
   @PostMapping("post/{postId}/comments")
   @CrossOrigin(origins = "http://localhost:4200")
-  private void insertComment(@PathVariable long postId, @RequestBody Comment comment, @AuthenticationPrincipal
-      UserDetails userDetails){
+  private ResponseEntity insertComment(@PathVariable long postId, @RequestBody Comment comment, @AuthenticationPrincipal
+      UserDetails userDetails) {
+    if (!postRepository.findById(postId).isPresent()) {
+      return ResponseEntity.badRequest().body("Oops! The post you're commenting on does not exist!");
+    }
     Post foundPost = postRepository.findById(postId).get();
+    if (comment.getCommentDescription().isEmpty()){
+      return ResponseEntity.badRequest().body("Oops! You can't post an empty comment!");
+    }
     comment.setPost(foundPost);
+    if (!accountRepository.findByUsername(userDetails.getUsername()).isPresent()) {
+      return ResponseEntity.badRequest().body("Something went wrong.");
+    }
     comment.setAccount(accountRepository.findByUsername(userDetails.getUsername()).get());
     commentRepository.save(comment);
+    return ok(comment);
   }
 
   @DeleteMapping("/{id}")
   @CrossOrigin(origins = "http://localhost:4200")
-  private void deleteComment(@PathVariable long commentId){
-    commentRepository.deleteById(commentId);
+  private ResponseEntity deleteComment(@PathVariable long commentId){
+    if (commentRepository.findById(commentId).isPresent()) {
+      commentRepository.deleteById(commentId);
+      return ok("Comment successfully deleted!");
+    } else {
+      return badRequest().body("Comment you're trying to delete does not exist.");
+    }
   }
 
 }
